+++ src/components/ProductListPage/ShopHeader.tsx
@@ -2,7 +2,7 @@ import styled from "styled-components";
 
 import { SCREEN_WIDTH_REM } from "@styles/GlobalStyle";
 import { ReactComponent as CartIcon } from "@assets/cart.svg";
-import { useCartItems } from "@src/contexts/cartItems/useCartItems";
+import { useCartItems } from "@src/hooks/query/useCartItems";
 
 const ShopHeader = () => {
   const { cartItems, isLoading } = useCartItems();
@@ -12,7 +12,7 @@ const ShopHeader = () => {
       <S.Title>SHOP</S.Title>
       <S.CartButton role="button" aria-label="장바구니 이동">
         <CartIcon />
-        {!isLoading && (
+        {cartItems && !isLoading && (
           <S.Badge>
             <S.CircleContent>{cartItems.length}</S.CircleContent>
           </S.Badge>
