--- src/contexts/cartItems/useCartItems.ts
@@ -1,12 +1,0 @@
-import { useContext } from "react";
-import { CartItemContextState, CartItemsContext } from "@contexts/cartItems/cartItemsContext";
-
-export const useCartItems = (): CartItemContextState => {
-  const context = useContext(CartItemsContext);
-
-  if (context === undefined) {
-    throw new Error("useCartItems는 CartItemsProvider 내부에서 사용되어야 합니다.");
-  }
-
-  return context;
-};
